feat: update load env, postgres, remove sqlite, new docs

+++ api/customer.go
@@ -170,8 +170,8 @@ func (server *Server) getCustomer(ctx *gin.Context) {
 // List Customer
 
 type listCustomerRequest struct {
-	PageID   int64 `form:"page_id" binding:"required,min=1"`
-	PageSize int64 `form:"page_size" binding:"required,min=5,max=10"`
+	PageID   int32 `form:"page_id" binding:"required,min=1"`
+	PageSize int32 `form:"page_size" binding:"required,min=5,max=10"`
 }
 
 func (server *Server) listCustomers(ctx *gin.Context) {
